Move handlers into subpackage
dispatcher/git_handler.py → dispatcher/handlers/git.py
dispatcher/gh_handler.py → dispatcher/handlers/gh.py

Removes redundant naming (everything in dispatcher/ is already dispatcher).
