+++ dispatcher/handlers/__init__.py
@@ -1,0 +1,1 @@
+"""Command handlers for git and gh CLI."""
